requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
supabase>=2.7.0
mailjet-rest>=1.3.4
google-generativeai>=0.3.0
aiofiles>=23.2.0
dnspython>=2.6.0